requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.7

# PDF Processing
PyMuPDF==1.23.8
Pillow==10.0.1

# LangChain & AI
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
openai==1.3.7
faiss-cpu==1.7.4

# NLP
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.0